change behavior how we decide if "changed" is true or false
- if we make a request to proxmox -> change=True otherwise false
- if apply=True and we have pending changes we request apply and changed=True

+++ plugins/modules/proxmox_sdn.py
@@ -23,30 +23,33 @@ sdn_object_id = re.compile(r"^[a-z][a-z0-9]*$")
 
 
 class ProxmoxSDNAnsible(ProxmoxAnsible):
-    def get_zone(self, zone):
+    def get_zone(self, zone, pending="0"):
         try:
-            zones = self.proxmox_api.cluster.sdn.zones.get()
-            return next((z for z in zones if z["zone"] == zone), None)
+            return next((z for z in self.proxmox_api.cluster.sdn.zones.get(pending=pending) if z["zone"] == zone), None)
         except Exception as e:
             self.module.fail_json(msg="Unable to retrieve zone: {0}".format(e))
 
-    def get_vnets_of_zone(self, zone):
-        return [vnet for vnet in self.proxmox_api.cluster.sdn.vnets.get() if vnet["zone"] == zone]
+    def get_vnets_of_zone(self, zone, pending="0"):
+        try:
+            return [vnet for vnet in self.proxmox_api.cluster.sdn.vnets.get(pending=pending) if vnet["zone"] == zone]
+        except Exception as e:
+            self.module.fail_json(msg="Unable to retrieve vnets: {0}".format(e))
 
     def create_zone(self, zone, update, apply):
         zone_exists = False
         if self.get_zone(zone["id"]):
             # Ugly ifs
-            if not update and not apply:
-                self.module.exit_json(changed=False, id=zone["id"], msg="sdn {0} already exists".format(zone["id"]))
-            if not update and apply:
+            if not update:
                 return False
+            # if not update and not apply:
+            #     return False
+            #     self.module.exit_json(changed=False, id=zone["id"], msg="sdn {0} already exists".format(zone["id"]))
+            # if not update and apply:
+            #     return False
             zone_exists = True
 
         zone_copy = dict(**zone)
         zone_id = zone_copy.pop("id")
-        if not sdn_object_id.match(zone_id):
-            self.module.fail_json(msg="{0} is not a valid sdn object identifier".format(zone_id))
 
         additionals = zone_copy.pop("additionals")
         if update:
@@ -60,23 +63,37 @@ class ProxmoxSDNAnsible(ProxmoxAnsible):
 
         return True
 
-    def delete_vnet(self, vnet):
-        if not self.get_vnet(vnet):
-            self.module.exit_json(changed=False, id=vnet, msg="vnet {0} doesn't exist".format(vnet))
+    def delete_vnet(self, force, vnet):
+        vnet_obj = self.get_vnet(vnet, pending="1")
+        if vnet_obj is None or vnet_obj.get("state") == "deleted":
+            return False
+
+        subnets = self.get_subnets_of_vnet(vnet)
+        if force and subnets:
+            for subnet in subnets:
+                self.delete_subnet(subnet["cidr"], vnet)
+        elif not force and subnets:
+            self.module.fail_json(
+                msg="Can't delete vnet {0} with subnets. Please remove subnets from vnet first or use `force: True`.".format(
+                    vnet
+                )
+            )
         try:
             self.proxmox_api.cluster.sdn.vnets(vnet).delete()
         except Exception as e:
             self.module.fail_json(msg="Failed to delete vnet with ID {0}: {1}".format(vnet, e))
+        return True
 
     def delete_zone(self, force, zone):
-        if not self.get_zone(zone):
-            self.module.exit_json(changed=False, id=zone, msg="Zone {0} doesn't exist".format(zone))
+        zone_obj = self.get_zone(zone, pending="1")
+        if zone_obj is None or zone_obj.get("state") == "deleted":
+            return False
 
         vnets = self.get_vnets_of_zone(zone)
 
         if force and vnets:
             for vnet in vnets:
-                self.delete_vnet(vnet["vnet"])
+                self.delete_vnet(force, vnet["vnet"])
         elif not force and vnets:
             self.module.fail_json(
                 msg="Can't delete zone {0} with vnets. Please remove vnets from zone first or use `force: True`.".format(
@@ -87,11 +104,11 @@ class ProxmoxSDNAnsible(ProxmoxAnsible):
             self.proxmox_api.cluster.sdn.zones(zone).delete()
         except Exception as e:
             self.module.fail_json(msg="Failed to delete zone with ID {0}: {1}".format(zone, e))
+        return True
 
-    def get_vnet(self, vnet):
+    def get_vnet(self, vnet, pending="0"):
         try:
-            vnets = self.proxmox_api.cluster.sdn.vnets.get()
-            return next((v for v in vnets if v["vnet"] == vnet), None)
+            return next((v for v in self.proxmox_api.cluster.sdn.vnets.get(pending=pending) if v["vnet"] == vnet), None)
         except Exception as e:
             self.module.fail_json(msg="Unable to retrieve vnet: {0}".format(e))
 
@@ -101,12 +118,10 @@ class ProxmoxSDNAnsible(ProxmoxAnsible):
         if self.get_vnet(vnet):
             # Ugly ifs
             if not update and not apply:
-                self.module.exit_json(changed=False, id=vnet, msg="Vnet {0} already exists".format(vnet))
+                return False
+                # self.module.exit_json(changed=False, id=vnet, msg="Vnet {0} already exists".format(vnet))
             if not update and apply:
                 return False
-
-        if not sdn_object_id.match(vnet):  # TODO should we do the check before creating zone
-            self.module.fail_json(msg="{0} is not a valid sdn object identifier".format(vnet))
 
         if update:
             self.proxmox_api.cluster.sdn.vnets(vnet).set(
@@ -119,20 +134,24 @@ class ProxmoxSDNAnsible(ProxmoxAnsible):
         return True
 
     def get_subnet(self, cidr, vnet):
+        return next((sdn for sdn in self.get_subnets_of_vnet(vnet) if sdn["cidr"] == cidr), None)
+
+    def get_subnets_of_vnet(self, vnet):
         try:
-            sdns = self.proxmox_api.cluster.sdn.vnets(vnet).subnets.get(pending="1")
-            return next((sdn for sdn in sdns if sdn["cidr"] == cidr), None)
+            return self.proxmox_api.cluster.sdn.vnets(vnet).subnets.get(pending="1")
         except Exception as e:
             self.module.fail_json(msg="Unable to retrieve subnet: {0}".format(e))
 
     def delete_subnet(self, subnet, vnet):
+        """Return `True` if subnet was deleted, `False` when already absent"""
         subnet_obj = self.get_subnet(subnet, vnet)
-        if not subnet_obj:
-            self.module.exit_json(changed=False, id=subnet, msg="Subnet {0} doesn't exist".format(subnet))
+        if subnet_obj is None or subnet_obj.get("state") == "deleted":
+            return False
         try:
             self.proxmox_api.cluster.sdn.vnets(vnet).subnets(subnet_obj["subnet"]).delete()
         except Exception as e:
             self.module.fail_json(msg="Failed to delete subnet with ID {0}: {1}".format(subnet, e))
+        return True
 
     def create_subnet(self, cidr, vnet, data, update=False, apply=False):
         vnet_object = self.get_vnet(vnet)
@@ -143,7 +162,8 @@ class ProxmoxSDNAnsible(ProxmoxAnsible):
         if subnet:
             # Ugly ifs
             if not update and not apply:
-                self.module.exit_json(changed=False, id=vnet, msg="Subnet {0} already exists".format(cidr))
+                return False
+                # self.module.exit_json(changed=False, id=vnet, msg="Subnet {0} does already exist".format(cidr))
             if not update and apply:
                 return False
 
@@ -232,6 +252,7 @@ def main():
     msgs = []
     pending_changes = False
 
+    # validation
     if subnet:
         cidr = subnet.pop("cidr")
         subnet_vnet = subnet.pop("vnet")
@@ -239,27 +260,35 @@ def main():
             _ = ipaddress.ip_network(cidr, False)
         except e:
             module.fail_json(msg=f"{cidr} not a valid CIDR")
+    if vnet:
+        if not sdn_object_id.match(vnet["id"]):
+            module.fail_json(msg="{0} is not a valid sdn object identifier".format(vnet["id"]))
 
-    if state == "present":
-        try:
+        if vnet["zone"] == "vlan" and not vnet.get("tag"):
+            module.fail_json(msg="missing vlan tag")
+    if zone:
+        zone_id = zone["id"]
+        if not sdn_object_id.match(zone_id):
+            module.fail_json(msg="{0} is not a valid sdn object identifier".format(zone_id))
+
+    # execute
+    try:
+        if state == "present":
             if zone:
-                check_changes = proxmox.create_zone(zone=zone, update=update, apply=apply)
-                if check_changes:
-                    # get zone info's to check for pending changes
-                    zone_info = proxmox.proxmox_api.cluster.sdn.zones(zone["id"]).get(pending="1")
-                    state = zone_info.get("state")
-                    if state:
-                        pending_changes = True
-                        msgs.append(
-                            "Zone {0} successfully {1}.".format(
-                                zone["id"], "changed" if state == "changed" else "created"
-                            )
-                        )
+                created_or_updated = proxmox.create_zone(zone=zone, update=update, apply=apply)
+                # if check_changes:
+                # get zone info's to check for pending changes
+                # zone_info = proxmox.proxmox_api.cluster.sdn.zones(zone_id).get(pending="1")
+                # zone_state = zone_info.get("state")
+                # if zone_state:
+                #     pending_changes = True
+                if created_or_updated:
+                    pending_changes = True
+                    msgs.append("Zone {0} successfully {1}.".format(zone["id"], "updated" if update else "created"))
+                else:
+                    msgs.append("Zone {0} already exists.".format(zone_id))
 
             if vnet:
-                if vnet["zone"] == "vlan" and not vnet.get("tag"):
-                    module.fail_json(msg="missing vlan tag")
-
                 check_changes = proxmox.create_vnet(
                     vnet=vnet["id"],
                     zone=vnet["zone"],
@@ -272,11 +301,14 @@ def main():
                 )
                 if check_changes:
                     # get vnet info's to check for pending changes
-                    vnet_info = proxmox.proxmox_api.cluster.sdn.vnets(vnet["id"]).get(pending="1")
-                    state = vnet_info.get("state", "")
-                    if state:
-                        pending_changes = True
-                        msgs.append("Vnet {0} successfully {1}.".format(vnet["id"], "updated" if update else "created"))
+                    # vnet_info = proxmox.proxmox_api.cluster.sdn.vnets(vnet["id"]).get(pending="1")
+                    # state = vnet_info.get("state")
+                    # if state:
+                    #     pending_changes = True
+                    pending_changes = True
+                    msgs.append("Vnet {0} successfully {1}.".format(vnet["id"], "updated" if update else "created"))
+                else:
+                    msgs.append("Vnet {0} already exists.".format(vnet["id"]))
 
             if subnet:
                 subnet["snat"] = ansible_to_proxmox_bool(subnet["snat"])
@@ -285,61 +317,88 @@ def main():
                 )
                 if check_changes:
                     # get subnet info's to check for pending changes
-                    subnet_info = proxmox.get_subnet(cidr, subnet_vnet)
-                    state = subnet_info.get("state")
-                    if state:
-                        pending_changes = True
-                        msgs.append("Subnet {0} successfully {1}.".format(cidr, "updated" if update else "created"))
-
-        except ResourceException as e:
-            if update:
-                module.fail_json(msg="Unable to update sdn objects. Reason: {0}".format(str(e)))
-            else:
-                module.fail_json(msg="Unable to create sdn objects. Reason: {0}".format(str(e)))
-        # if we have pending changes, we have changed something
-        data["changed"] = pending_changes
-        if pending_changes and apply:
-            proxmox.apply_changes()
-            msgs.append("Pending changes applied.")
-        if not pending_changes:
-            msgs.append("Everything is up to date.")
-        module.exit_json(applied=apply, **data, msg="\n".join(msgs))
-
-    else:
-        if subnet:
-            proxmox.delete_subnet(cidr, subnet_vnet)
-            subnet_info = proxmox.get_subnet(cidr, subnet_vnet)
-            if subnet_info:  # can be none if subnet was in 'new' state before and is now deleted
-                state = subnet_info.get("state")
-                if state:
+                    # subnet_info = proxmox.get_subnet(cidr, subnet_vnet)
+                    # state = subnet_info.get("state")
+                    # if state:
+                    #     pending_changes = True
                     pending_changes = True
-                    msgs.append("Subnet {0} deleted".format(cidr))
+                    msgs.append("Subnet {0} successfully {1}.".format(cidr, "updated" if update else "created"))
+                else:
+                    msgs.append("Subnet {0} of {1} already exists.".format(cidr, subnet_vnet))
 
-        if vnet:
-            proxmox.delete_vnet(vnet["id"])
-            # get vnet info's to check for pending changes
-            vnet_info = proxmox.proxmox_api.cluster.sdn.vnets(vnet["id"]).get(pending="1")
-            state = vnet_info.get("state")
-            if state:
-                pending_changes = True
-                msgs.append("Vnet {0} deleted".format(vnet["id"]))
+        else:  # delete objects
+            if subnet:
+                check_changes = proxmox.delete_subnet(cidr, subnet_vnet)
+                if not check_changes:
+                    msgs.append("Subnet '{0}' of vnet {1} is already absent".format(cidr, subnet_vnet))
+                else:
+                    # subnet_info = proxmox.get_subnet(cidr, subnet_vnet)
+                    # if subnet_info:  # can be none if subnet was in 'new' state before and is now deleted
+                    #     state = subnet_info.get("state")
+                    #     if state:
+                    pending_changes = True
+                    msgs.append("Subnet {0} of {1} deleted.".format(cidr, subnet_vnet))
 
-        if zone:
-            proxmox.delete_zone(module.params["force"], zone["id"])
-            # get zone info's to check for pending changes
-            zone_info = proxmox.proxmox_api.cluster.sdn.zones(zone["id"]).get(pending="1")
-            state = zone_info.get("state")
-            if state:
-                pending_changes = True
-                msgs.append("Zone {0} deleted".format(zone["id"]))
+            if vnet:
+                check_changes_vnet = proxmox.delete_vnet(module.params["force"], vnet["id"])
+                if not check_changes_vnet:
+                    msgs.append("Vnet '{0}' is already absent".format(vnet["id"]))
+                else:
+                    # get vnet info's to check for pending changes
+                    # vnet_info = proxmox.proxmox_api.cluster.sdn.vnets(vnet["id"]).get(pending="1")
+                    # state = vnet_info.get("state")
+                    # if state:
+                    pending_changes = True
+                    msgs.append("Vnet {0} deleted.".format(vnet["id"]))
+
+            if zone:
+                check_changes = proxmox.delete_zone(module.params["force"], zone_id)
+                if not check_changes:
+                    msgs.append("Zone '{0}' is already absent".format(zone_id))
+                else:
+                    # get zone info's to check for pending changes
+                    # zone_info = proxmox.proxmox_api.cluster.sdn.zones(zone_id).get(pending="1")
+                    # state = zone_info.get("state")
+                    # if state:
+                    pending_changes = True
+                    msgs.append("Zone {0} deleted.".format(zone_id))
+
+        # if we have pending changes, we have changed something
+        # if pending_changes:
+        if apply:
+            if not pending_changes:
+                zone_info = proxmox.proxmox_api.cluster.sdn.zones.get(pending="1")
+                pending_changes = any(info.get("state") for info in zone_info)
+            if not pending_changes:
+                vnet_info = proxmox.proxmox_api.cluster.sdn.vnets.get(pending="1")
+                pending_changes = any(info.get("state") for info in vnet_info)
+            if not pending_changes:
+                for info in vnet_info:
+                    subnet_info = proxmox.get_subnets_of_vnet(info["vnet"])
+                    pending_changes = any(sinfo.get("state") for sinfo in subnet_info)
+                    if pending_changes:
+                        break
+            if pending_changes:
+                proxmox.apply_changes()
+                msgs.append("Pending changes applied.")
+            else:
+                # msgs.append(zone_info)
+                msgs.append("No pending changes detected.")
+        # else:
+        #     msgs.append("Pending changes not applied.")
         data["changed"] = pending_changes
-        if pending_changes and apply:
-            proxmox.apply_changes()
-            msgs.append("Pending changes applied.")
-        if not pending_changes:
-            msgs.append("Everything is up to date.")
+        # else:
+        #     msgs.append("Everything is up to date.")
 
-        module.exit_json(applied=apply, **data, msg="\n".join(msgs))
+        module.exit_json(**data, msg=" ".join(msgs))
+
+    except ResourceException as e:
+        if update:
+            module.fail_json(msg="Unable to update sdn objects. Reason: {0}".format(str(e)))
+        elif state == "absent":
+            module.fail_json(msg="Unable to delete sdn objects. Reason: {0}".format(str(e)))
+        else:
+            module.fail_json(msg="Unable to create sdn objects. Reason: {0}".format(str(e)))
 
 
 if __name__ == "__main__":
